packages/core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rpaforge-core"
version = "0.1.0.dev1"
description = "Core engine for RPAForge - Robot Framework wrapper with debugging and recording"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "RPAForge Contributors"}
]
keywords = [
    "rpa",
    "automation",
    "robot-framework",
    "debugger",
    "recorder",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Robot Framework",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "robotframework>=7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/chelslava/rpaforge"
Repository = "https://github.com/chelslava/rpaforge.git"
Issues = "https://github.com/chelslava/rpaforge/issues"

[tool.setuptools.packages.find]
where = ["src"]
